nexus_py_sdk/requirements.txt
#Core dependencies for the Nexus SDK

cryptography>=43.0.0
pydantic>=2.0.0
requests>=2.31.0

#Optional: For Google Secret Manager support (Prod mode)

google-cloud-secret-manager>=2.16.0